STEP1 기반 다지기 - Combatant와 체간 시스템 구현 내용 목차
- 추가된 컨텐츠: Enemy가 플레이어의 위치로 이동하도록 변경.

[미해결 사항]
- 현재 게이지 바가 문자가 깨져서 나오며, 체간이 바뀌었을 때만 생겼다 사라지는 문제가 있다. => 체간 게이지 문자열 표시에서 배경색 변경 방식으로 수정해야함.
- Enemy와의 충돌 판정도 넣어야 할 듯.

+++ Engine/Math/Vector2.cpp
@@ -12,7 +12,7 @@ namespace Wanted
 	{
 	}
 
-	Vector2::Vector2(float x, float y)
+	Vector2::Vector2(int x, int y)
 		: x(x), y(y)
 	{
 	}
